Extracted HttpClient from OnlineHubService
DAT-4800
