+++ liquibase-core/src/main/java/liquibase/hub/core/OnlineHubService.java
@@ -42,23 +42,18 @@ import java.text.SimpleDateFormat;
 import java.util.*;
 
 public class OnlineHubService implements HubService {
-    private static final Logger LOG = Scope.getCurrentScope().getLog(OnlineHubService.class);
-    private Yaml yaml;
+
     private UUID organizationId;
     private Boolean hasApiKey;
 
+    private HttpClient http;
+    
     public OnlineHubService() {
-        DumperOptions dumperOptions = new DumperOptions();
+        this.http = createHttpClient();
+    }
 
-        dumperOptions.setPrettyFlow(true);
-        dumperOptions.setDefaultFlowStyle(DumperOptions.FlowStyle.FLOW);
-        dumperOptions.setDefaultScalarStyle(DumperOptions.ScalarStyle.DOUBLE_QUOTED);
-        dumperOptions.setWidth(Integer.MAX_VALUE);
-
-        yaml = new Yaml(new Constructor(), new HubRepresenter(), dumperOptions);
-
-        yaml.setBeanAccess(BeanAccess.FIELD);
-
+    public HttpClient createHttpClient() {
+        return new HttpClient();
     }
 
     @Override
@@ -77,7 +72,7 @@ public class OnlineHubService implements HubService {
 
     @Override
     public HubUser getMe() throws LiquibaseHubException {
-        final Map response = doGet("/api/v1/users/me", Map.class);
+        final Map response = http.doGet("/api/v1/users/me", Map.class);
 
         HubUser user = new HubUser();
         user.setId(UUID.fromString((String) response.get("id")));
@@ -88,7 +83,7 @@ public class OnlineHubService implements HubService {
 
     @Override
     public Organization getOrganization() throws LiquibaseHubException {
-        final Map<String, List<Map>> response = doGet("/api/v1/organizations", Map.class);
+        final Map<String, List<Map>> response = http.doGet("/api/v1/organizations", Map.class);
 
         Organization org = new Organization();
         List<Map> contentList = response.get("content");
@@ -109,7 +104,7 @@ public class OnlineHubService implements HubService {
     public List<Project> getProjects() throws LiquibaseHubException {
         final UUID organizationId = getOrganization().getId();
 
-        final Map<String, List<Map>> response = doGet("/api/v1/organizations/" + organizationId.toString() + "/projects", Map.class);
+        final Map<String, List<Map>> response = http.doGet("/api/v1/organizations/" + organizationId.toString() + "/projects", Map.class);
         List<Map> contentList = response.get("content");
         List<Project> returnList = new ArrayList<>();
         for (int i = 0; i < contentList.size(); i++) {
@@ -121,7 +116,7 @@ public class OnlineHubService implements HubService {
                 date = parseDate(dateString);
             }
             catch (ParseException dpe) {
-                LOG.warning("Project '" + name + "' has an invalid create date of '" + dateString + "'");
+                Scope.getCurrentScope().getLog(getClass()).warning("Project '" + name + "' has an invalid create date of '" + dateString + "'");
             }
             Project project = new Project();
             project.setId(UUID.fromString(id));
@@ -138,7 +133,7 @@ public class OnlineHubService implements HubService {
         final UUID organizationId = getOrganization().getId();
         Project project = new Project();
         project.setName(projectName);
-        return doPost("/api/v1/organizations/" + organizationId.toString() + "/projects", project, Project.class);
+        return http.doPost("/api/v1/organizations/" + organizationId.toString() + "/projects", project, Project.class);
     }
 
     @Override
@@ -148,7 +143,7 @@ public class OnlineHubService implements HubService {
         changeLogRequest.setExternalChangelogId(UUID.randomUUID().toString());
         changeLogRequest.setFileName("string");
         changeLogRequest.setName("string");
-        return doPost("/api/v1/organizations/" + organizationId.toString() + "/projects/" + project.getId() + "/changelogs", changeLogRequest, HubChangeLog.class);
+        return http.doPost("/api/v1/organizations/" + organizationId.toString() + "/projects/" + project.getId() + "/changelogs", changeLogRequest, HubChangeLog.class);
     }
 
     @Override
@@ -167,7 +162,7 @@ public class OnlineHubService implements HubService {
     public Environment getEnvironment(Environment exampleEnvironment, boolean createIfNotExists) throws LiquibaseHubException {
         if (exampleEnvironment.getId() != null) {
             //do not auto-create if specifying the exact id
-            return doGet("/api/v1/environments/" + exampleEnvironment.getId().toString(), null, Environment.class);
+            return http.doGet("/api/v1/environments/" + exampleEnvironment.getId().toString(), null, Environment.class);
         }
 
         final List<Environment> environments;
@@ -198,7 +193,7 @@ public class OnlineHubService implements HubService {
     public List<Environment> getEnvironments(Environment exampleEnvironment) throws LiquibaseHubException {
         final Organization organization = getOrganization();
 
-        final Map response = doGet("/api/v1/organizations/" + organization.getId() + "/environments", Collections.singletonMap("search", toSearchString(exampleEnvironment)), Map.class);
+        final Map response = http.doGet("/api/v1/organizations/" + organization.getId() + "/environments", Collections.singletonMap("search", toSearchString(exampleEnvironment)), Map.class);
 
         List<Environment> returnList = new ArrayList<>();
 
@@ -233,7 +228,7 @@ public class OnlineHubService implements HubService {
         if (environment.getProject() == null || environment.getProject().getId() == null) {
             throw new LiquibaseHubUserException("projetId is required to create an environment");
         }
-        return doPost("/api/v1/organizations/" + getOrganization().getId() + "/projects/" + environment.getProject().getId() + "/environments", environment, Environment.class);
+        return http.doPost("/api/v1/organizations/" + getOrganization().getId() + "/projects/" + environment.getProject().getId() + "/environments", environment, Environment.class);
     }
 
     /**
@@ -252,7 +247,7 @@ public class OnlineHubService implements HubService {
         for (Project project : projects) {
             try {
                 Map<String, String> response =
-                    doGet("/api/v1/organizations/" + organizationId.toString() + "/projects/" + project.getId() + "/changelogs/" + changeLogId, Map.class);
+                    http.doGet("/api/v1/organizations/" + organizationId.toString() + "/projects/" + project.getId() + "/changelogs/" + changeLogId, Map.class);
                 HubChangeLog hubChangeLog = createHubChangeLogFromResponse(response);
                 hubChangeLog.setProject(project);
                 return hubChangeLog;
@@ -284,102 +279,6 @@ public class OnlineHubService implements HubService {
         return hubChangeLog;
     }
 
-    protected <T> T doGet(String url, Class<T> returnType) throws LiquibaseHubException {
-        return doGet(url, new HashMap<>(), returnType);
-
-    }
-
-    protected <T> T doGet(String url, Map<String, String> parameters, Class<T> returnType) throws LiquibaseHubException {
-        if (parameters != null && parameters.size() > 0) {
-            url += "?" + toQueryString(parameters);
-        }
-
-        URLConnection connection = openConnection(url);
-
-        return doRequest(connection, returnType);
-    }
-
-    protected <T> T doPost(String url, Object requestBodyObject, Class<T> returnType) throws LiquibaseHubException {
-        HttpURLConnection connection = (HttpURLConnection) openConnection(url);
-        connection.setDoOutput(true);
-
-        String requestBody = createPostRequestBody(requestBodyObject);
-        try {
-            try (OutputStream output = connection.getOutputStream()) {
-                output.write(requestBody.getBytes(StandardCharsets.UTF_8));
-            }
-            return doRequest(connection, returnType);
-        } catch (IOException e) {
-            throw new LiquibaseHubException(e);
-        }
-    }
-
-    private String createPostRequestBody(Object requestObjectBody) {
-        return yaml.dumpAs(requestObjectBody, Tag.MAP, DumperOptions.FlowStyle.FLOW);
-    }
-
-    private URLConnection openConnection(String url) throws LiquibaseHubException {
-        HubConfiguration hubConfiguration = LiquibaseConfiguration.getInstance().getConfiguration(HubConfiguration.class);
-        String hubUrl = hubConfiguration.getLiquibaseHubUrl();
-        String apiKey = hubConfiguration.getLiquibaseHubApiKey();
-
-        try {
-            final URLConnection connection = new URL(hubUrl + url).openConnection();
-            connection.setRequestProperty("User-Agent", "Liquibase " + LiquibaseUtil.getBuildVersion());
-            connection.setRequestProperty("Authorization", "Bearer " + apiKey);
-            connection.setRequestProperty("Accept", "application/json");
-            connection.setRequestProperty("Content-Type", "application/json");
-            return connection;
-        } catch (IOException e) {
-            throw new LiquibaseHubException(e);
-        }
-    }
-
-    private String toQueryString(Map<String, String> parameters) {
-        if (parameters == null) {
-            return "";
-        }
-
-        List<String> paramArray = new ArrayList<>();
-        try {
-            for (Map.Entry<String, String> entry : parameters.entrySet()) {
-                paramArray.add(entry.getKey() + "=" + URLEncoder.encode(entry.getValue(), "UTF-8"));
-            }
-        } catch (UnsupportedEncodingException e) {
-            throw new RuntimeException(e);
-        }
-
-        return StringUtil.join(paramArray, "&");
-    }
-
-    private <T> T doRequest(URLConnection connection, Class<T> returnType) throws LiquibaseHubException {
-        try (InputStream response = connection.getInputStream()) {
-            return (T) yaml.loadAs(response, returnType);
-        } catch (IOException e) {
-            try {
-                try (InputStream error = ((HttpURLConnection) connection).getErrorStream()) {
-                    if (error != null) {
-                        final Map errorDetails = yaml.load(error);
-
-                        LiquibaseHubException returnException = new LiquibaseHubException((String) errorDetails.get("message"), e);
-
-                        if (((HttpURLConnection) connection).getResponseCode() == 404) {
-                            returnException = new LiquibaseHubObjectNotFoundException(returnException.getMessage(), returnException.getCause());
-                        }
-                        returnException.setTimestamp((String) errorDetails.get("timestamp"));
-                        returnException.setDetails((String) errorDetails.get("details"));
-                        throw returnException;
-
-                    }
-                }
-            } catch (IOException ioException) {
-                Scope.getCurrentScope().getLog(getClass()).info("Cannot read request error stream", e);
-            }
-
-            throw new LiquibaseHubException(e);
-        }
-    }
-
     /**
      * Converts an object to a search string. Any properties with non-null values are used as search arguments
      */
@@ -408,18 +307,4 @@ public class OnlineHubService implements HubService {
 
     }
 
-    private static class HubRepresenter extends Representer {
-
-        public HubRepresenter() {
-            getPropertyUtils().setSkipMissingProperties(true);
-        }
-
-        @Override
-        protected NodeTuple representJavaBeanProperty(Object javaBean, Property property, Object propertyValue, Tag customTag) {
-            if (propertyValue == null) {
-                return null;
-            }
-            return super.representJavaBeanProperty(javaBean, property, propertyValue, customTag);
-        }
-    }
 }
